MVTools: Move typed dicts to typing submodule
Also replaces the typing in `MVTools.__init__`. I saw no real reason not to reuse these where possible.
