+++ vsdenoise/mvtools/__init__.py
@@ -6,4 +6,5 @@ from .enums import *
 from .motion import *
 from .mvtools import *
 from .presets import *
+from .types import *
 from .utils import *
